Add and use "pure" functions and expressions
Pure functions are deterministic, context-independent and
focus-independent. File fn/pure.rnc lists them according to the
xpath-3.0 reference. Purity depends on the arity of the function.

Use pure functions to define pure expressions, built from literals,
variables and pure functions (with the correct arity when it matters).
This is used in place of literals in xpath-1.0-navigational. Note
that, for this file, pure functions should be restricted to pure 1.0
functions; that is left for a future commit.